Refactor background
