+++ src/containers/Elements/index.js
@@ -9,4 +9,5 @@ export { default as Sociogram } from './Sociogram';
 export { default as EdgeLayout } from './EdgeLayout';
 export { default as NodeLayout } from './NodeLayout';
 export { default as NodeBucket } from './NodeBucket';
+export { default as SociogramRadar } from './SociogramRadar';
 export { default as SociogramBackground } from './SociogramBackground';
