+++ src/styles/components/_all.scss
@@ -18,5 +18,6 @@
 @import 'interface';
 @import 'sociogram';
 @import 'sociogram-background';
+@import 'sociogram-radar';
 @import 'node-layout';
 @import 'node-bucket';
