+++ src/styles/components/_node-layout.scss
@@ -11,6 +11,9 @@
 
     .node {
       margin: 0;
+      // TODO: Add size presets to UI?
+      width: 100px;
+      height: 100px;
     }
   }
 }
