+++ src/styles/components/_sociogram-background.scss
@@ -4,8 +4,4 @@
   height: 100%;
   top: 0;
   left: 0;
-
-  &__range {
-    fill: transparentize(palette('graph-data-1'), .3);
-  }
 }
