+++ src/styles/components/_sociogram-radar.scss
@@ -1,0 +1,11 @@
+.sociogram-radar {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+
+  &__range {
+    fill: transparentize(palette('graph-data-1'), .4);
+  }
+}
